added interface/address exclusion support to watchlink. The correct format is:
[interface]
[interface] [ip/mask]

It should also accept a plain IP--but I haven't tested this yet.

Also added syslogging at the info level on interface up/down events.

This change provides support for HA to work with interface cable event. HA and VRRP will need to
populate the /var/linkstatus/exclude file with the VIPs under their management.

In addition to circumvent an additional problem in HA there needs to be an additional address on
the monitored interface, otherwise HA will admin down the interface when the cable is pulled