modified watchlink parser to support the following behaviors:
> comments are now allowed after the # symbol after each entry
> address ranges are now supported in the matching if the user
enters a netmask value, then every address within the range is
exlcluded
> ip value (without a mask) is treated as a /32 for matching
> matching is against the masked off address value, the mask
is specified by the exclusion entry (and not by the configured
address value)
> sig_usr1 will reload the configuration and is required after
each modification of the exclude file.

The format of the file is the following:

[interface] [address|network] #comments

Only one entry per line is allowed

vifs are supported in the interface line (i.e. eth0.1w)