Two passes on update (ipv4 and ipv6)
When interface comes down, need two passes over the rib to correctly
process rib updates.  If interface has both ipv6 and ipv4 routes and
comes down, need to do it as a two step process. Not sure exactly why,
but suspect it has to do with the the work queue and the flagging of
activity to the head of the list.