Revised version if interface renaming
This resolves issue with static routes of multilink.
In the multilink case, route is created before device exists.
Later device is created as ppp0 and gets renamed to ml0.

Zebra needed to orphan the old interface structure and create
a new one to correctly match up the interface with the inactive
route.