Fix BGP IPV4 in IPV6 support
By default on Linux, if an IPV6 socket is available for a port,
it will receive incoming IPV4 connections (as IPV4-in-IPV6 address),
unless the socket is explictly marked as IPV6 only.  BGP was
receiving the incomming connection but rejecting it because the
source address didn't match the peer.

Fix was to force IPV6_only.