BGP: fix crash when allowas-in is done
This fixes BGP crash when allow-as is changed on inactive peer.
It supersedes Michael Larson's earlier patch.

When allowas-in is changed on a peer that is not up, BGP would crash
trying to do route_refresh.  If peer is not up, there is no need
to do notification or send.