lib: get rid of extraneous output with vty_prepend
the vty_*prepend* family can be used to reduce the amount of output
produced by "show running-config" and "write ...". it buffers output in
struct vty->prepend (1024 bytes) and outputs it when vty_out is called.
if vty_out isn't called, it can be removed with vty_unprepend later.

applied on zebra and ospfd to get rid of empty interface blocks.