ospfd: SPF changes for unnumbered interfaces
Modify the SPF calculation based on Paul Jakma's work at Quagga

http://code.quagga.net/?p=people/paul/quagga;a=commitdiff;h=c705c1b07fed02cacc5b
213ebd8be64d5c0c1920

The gist of the change is pretty simple... when we start off with us
as the reference vertex, we need to check to see if we encounter an
unnumbered interface which is defined by the LSA link_data field
having 0 in it's upper byte (i.e. not a valid IP address).  In this
case, we need to find the ospf_interface by the ifindex as opposed to
the existing code which determines the nexthop router's link IP
address by examining it's LSA links.