ospfd/zebra: get host routes to peers into kernel
Get host routes to peers into the kernel routing table

ospfd and zebra expect that all routes go through peers that are
adjacent through an attached subnet (i.e. 10.2.3.0/24 via 192.168.0.1
where one of our interfaces has an address on the 192.168.0.0/x
subnet).

This patch allows us to direct routes through a host route to an
adjacent peer which is installed pointing to an interface (ifindex);
i.e. a host route.  This is equivalent to the following manual
configuration.

  ip route add 10.3.0.1 dev eth1
  ip route add 10.2.3.0/24 via 10.3.0.1

The linux generous ARP receive characteristic insures that 10.3.0.1
answers an ARP on our eth1.