ospfd: refine Router-LSA "# TOS" length check
ospf_router_lsa_links_examin() was checking structural integrity of a
Router-LSA link descriptors list without checking, that "# TOS" field of
the current descriptor actually belongs to the space of the received
packet. Although, considering buffer size, this could not cause any
issues, this commit makes the check consistent in this regard.