zebra: fix misplaced CMD_ERR_NO_MATCH in rtadv CLI
CMD_ERR_NO_MATCH is used to indicate a command, which does not exist. In
the case of no_ipv6_nd_rdnss_addr() and no_ipv6_nd_dnssl_domain() the
command exists, but the item in question does not. An error message and
CMD_WARNING error code is the right response in such cases.