isisd: initialize metric-style as narrow
Let isisd metric-style default to narrow so it actually implements
the behaviour isis_config_write() expects.

This allows a narrow metric style to be safed. Before this patch, that
was not possible because isis_config_write() does not produce a
statement for narrow metric style, assuming it is the default. Once
isisd was restarted, the metric style would then default back to wide.

Signed-off-by: Christian Franke <chris@opensourcerouting.org>