[OSRQUAGGA-59] Quagga does not timeout during BGP Open
OpenSent is the only FSM state where HoldTimer MUST be set,
and its value can not be negotiated or configured, and its recommended
value is 4 minutes.

The configured HoldTime value (default 90 sec) is the one offered to peer,
and is negotiated during the OpenSent state, and is first used in
OpenConfirm state.

The previous incorrect behavior was to use the configured HoldTime as the
value for HoldTimer in OpenSent state (which the user could set to 0,
and skip enabling the timer in the OpenSent state when it is mandatory.

Moved the correct code to re-set the timers from "stop" to "connect_successful"
function, so the negotiated values are sure to be reverted on a new
connection.