HoldTimer is mandatory in OpenSent state, revert to configured HoldTIme when disconnected
Two related issues are addressed:
OpenSent is the only FSM state where HoldTimer MUST be set, and its value can not be
negotiated or configured, and its recommended value is 4 minutes.

The configured HoldTime value (default 90 sec) is the one offered to peer,
and is negotiated during the OpenSent state, and is first used in OpenConfirm state.

WHen a peer connection is lost (e.g. when the peer closes or times out), the HoldTime
value that will be offered at the next negotiation to the same peer must be reset to the
value configured by the user, and should not retain the previously negotiated value.

The previous incorrect behavior was to use the configured HoldTime as the value for
HoldTimer in OpenSent state (which the user could set to 0, or >= 3 sec), and to retain
the negotiated HoldTime and then re-use it in subsequent connect attampt with same peer
during OpenSent state, and as initial value for negotiation.

Fixes the following issues in OSR JIRA:
[OSRQUAGGA-59] ANVL BGP4-21.5 Quagga does not timeout during BGP Open