zebra: warn if advertising connected with _REAL unset
The implementation in zebra and the zclient protocol allow to communicate
addresses to clients which are not yet in the kernel.

This is usually not done and most clients seem to expect an address to be
configured in the kernel when they receive it. Therefore, it seems
reasonable to issue a warning when advertising an address to the clients
that is not yet in the kernel.

Signed-off-by: Christian Franke <chris@opensourcerouting.org>